Add files via upload
node: This class represents a node in the doubly linked list. It contains three attributes: "data" to store the value of the node, "prv" to store a reference to the previous node, and "nx" to store a reference to the next node.

ll: This class represents the doubly linked list itself. It has methods to perform various operations on the list:

 __init__: Initializes the linked list with a head pointer set to None.
  
  insertf: Inserts a node at the beginning of the list.
  
  inserte: Inserts a node at the end of the list.
  
  insertm: Inserts a node after a given node with specific data.
  
  display: Displays the elements of the list.
  
  deletef: Deletes the first node of the list.
  
  deletee: Deletes the last node of the list.
  
  deletem: Deletes a specific node with given data.
  
The code then provides a simple user interface to interact with the doubly linked list through a while loop, where the user can choose operations like inserting, deleting, or displaying elements of the list until choosing to exit.