Device
This Python code implements a singly linked list with various operations. Here's a brief description of each class and method:

node: This class represents a node in the singly linked list. 
It has two attributes:
             "d" to store the data of the node and "n" to store the reference to the next node.

listop: This class represents the operations that can be performed on the linked list.
__init__: Initializes the linked list with a head pointer set to None.
insertats: Inserts a node at the beginning of the list.
insertafter: Inserts a node after a specified node.
insertate: Inserts a node at the end of the list.
traverse: Traverses the linked list and prints the data of each node.
deleteatf: Deletes the first node of the list.
deleteatl: Deletes the last node of the list.
deletem: Deletes a specific node with given data.
  
The code then provides a simple user interface to interact with the linked list through a while loop. The user can choose operations like inserting, deleting, or displaying elements of the list until choosing to exit.