device
This Python code implements a circular linked list with the following functionalities:

node: This class represents a node in the circular linked list. It has two attributes: `data` to store the value of the node and `nxt` to store a reference to the next node.
ll: This class represents the circular linked list and contains methods to perform various operations on the list.
__init__: Initializes the circular linked list with both `head` and `tail` set to None.
insertf: Inserts a node at the beginning of the list.
inserte: Inserts a node at the end of the list.
insertm: Inserts a node after a specified value.
deletef: Deletes the first node of the list.
deletee: Deletes the last node of the list.
deletem: Deletes a specific node with given data.
display: Displays the elements of the circular linked list.
  
The code then provides a user interface to interact with the circular linked list through a while loop. The user can choose operations like inserting, deleting, or displaying elements of the list until choosing to exit.